Cleaned scripts, removed comments, moved all functions to utils library, deleted files which are not needed, deleted lammps inputs which are not used
