--- mother+daughter.in
@@ -1,38 +1,0 @@
-
-
-units		    lj
-dimension	    3
-atom_style      full 
-neigh_modify	delay 0 every 1
-boundary		p p p
-
-include ./sim_params.lmp   #include the file with all parameters of the simulation
-
-timestep ${Integration_step}
-
-
-read_restart       thermalization.rst
-
-variable vel_var equal (${T}/2)/1000
-
-variable srand equal ${perturbation_seed}
-
-variable rand_v atom normal(0,v_vel_var,v_srand)
-
-variable        vx atom vx+v_rand_v
-variable        vy atom vy+v_rand_v
-variable        vz atom vz+v_rand_v
-
-
-set             atom * vx v_vx 
-set             atom * vy v_vy 
-set             atom * vz v_vz
-
-
-velocity all scale ${T}
-velocity all zero linear
-velocity all zero angular
-
-fix             NVT_decorrelation all nvt temp ${T} ${T} ${Thermo_damp} tchain 1
-  
-
