--- mother.in
@@ -1,60 +1,0 @@
-#####MOTHER TRAJECTORY WHERE I GENERATE ALL THE INITIAL STATES FOR THE NONEQUILIBRIUM DAUGHTER SIMULATIONS ######
-
-processors 1 1 1
-
-units		    lj
-dimension	    3
-atom_style      full 
-neigh_modify	delay 0 every 1
-boundary		p p p
-
-include ./sim_params.lmp   #include the file with all parameters of the simulation
-
-timestep ${Integration_step}
-
-lattice         fcc ${rho_f}
-region          simbox prism 0 ${L} 0 ${L} 0 ${L} 0 0 0 units box
-create_box      1 simbox 
-
-create_atoms    1 region simbox
-
-group           fluid region simbox
-
-set             group fluid type 1
-
-mass            * 1.0
-pair_style      lj/cut ${rc}
-
-pair_coeff       1 1 1.0 1.0
-
-velocity        fluid create $T ${rseed}
-
-thermo 1000
-
-fix             NVT_thermalization all nvt temp ${T} ${T} ${Thermo_damp} tchain 1    #classic Nose Hoover dynamics
-
-run   ${Nthermalization}
-
-variable	Pxy equal pxy
-
-fix   	pxy all ave/time 1 1 ${Ninloop} v_Pxy ave one file Mother_Output.txt format %15.10g mode scalar  #check to verify whether the pressure at t=0 for the nonequilibrium trajectory matches the equilibrium one  
-
-
-##In the following is generate the daugher starting points throught a loop cycle 
-
-label loop
-variable index loop   ${Nchild}
-
-	run   ${Ninloop}
-	write_data    branch${index}.dat pair ij
-    
-next index
-jump SELF loop
-
-
-
-
-
-
-
-
